better image-source implementation, not in terms of pressure though

+++ raytracer/src/attenuator_program.cpp
@@ -78,7 +78,7 @@ kernel void hrtf_kernel(float3 mic_pos,
                  global volume_type * hrtfData,
                  float3 pointing,
                  float3 up,
-                 ulong channel) {
+                 uint channel) {
     size_t i = get_global_id(0);
     const float WIDTH = 0.1;
